+++ modules/preview-react/loading-sparkles/lib/sparkleIcon.tsx
