+++ modules/preview-react/loading-sparkles/stories/LoadingSparkles.stories.mdx
@@ -1,10 +1,10 @@
-import {LoadingSparkles} from '@workday/canvas-kit-react/loading-sparkles';
+import {LoadingSparkles} from '@workday/canvas-kit-preview-react/loading-sparkles';
 
 import {Basic} from './examples/Basic';
 import {RTL} from './examples/RTL';
 import {Accessible} from './examples/Accessible';
 
-<Meta title="Components/Indicators/Loading Sparkles" component={LoadingSparkles} />
+<Meta title="Preview/Loading Sparkles" component={LoadingSparkles} />
 
 # Canvas Kit Loading Sparkles
