+++ modules/preview-react/loading-sparkles/stories/examples/Accessible.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import {LoadingSparkles} from '@workday/canvas-kit-react/loading-sparkles';
+import {LoadingSparkles} from '@workday/canvas-kit-preview-react/loading-sparkles';
 
 export const Accessible = () => {
   return <LoadingSparkles aria-live="polite" aria-label="Loading Users" />;
