+++ modules/preview-react/loading-sparkles/stories/examples/RTL.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import {LoadingSparkles} from '@workday/canvas-kit-react/loading-sparkles';
+import {LoadingSparkles} from '@workday/canvas-kit-preview-react/loading-sparkles';
 import {CanvasProvider, ContentDirection} from '@workday/canvas-kit-react/common';
 
 export const RTL = () => {
